Add files via upload
This is the file I displayed in class today to get the dates read and processed. This should get you started. Please try to get the entire program written (all URLs generated, chromedriver working, cheapest price scraped (will need to wait for it to load using time.sleep(10) or something similar. create pandas data frame for all parameters, cheapest fare and urls. Work with others if needed, but please come to class with a working version.